pipeline/requirements.txt
pyyaml
requests
name2nat
ethnicolr
pandas
tqdm
